some bugfixes in the queue handling of the ACF multi thread framework
git-svn-id: http://astade.tigris.org/svn/astade/trunk@2347 aaa30fad-34f9-0310-8aa9-877cc762ee2e

+++ Model/Package_Astade/Package_Astade_framework/Package_HQHTBXGO/classes/class_WOCZHMNZ/operations/operation_HFIITVZA/code.cpp
@@ -1,2 +1,2 @@
 //~~ int getShortestTimer() [ACF] ~~
-return scheduledTimeouts ? my_Timeouts[0].Time : -1;
+return scheduledTimeouts ? my_Timeouts[0].Time : (unsigned int)-1;
